[api] Support v6.1 (#226)
https://core.telegram.org/bots/api#june-20-2022

Bot API 6.1

Join Requests & Payments

  - [x]  Added the fields join_to_send_messages and join_by_request to the class [Chat](https://core.telegram.org/bots/api#chat).
  - [x]  Added the ability to process join requests which were created [without an invite link](https://telegram.org/blog/700-million-and-premium#join-requests-for-public-groups). Bots will receive a “chat_join_request” update as usual.
  - [x]  Added the method [createInvoiceLink](https://core.telegram.org/bots/api#createinvoicelink) to generate an HTTP link for an invoice.

Telegram Premium Support ([more info](https://telegram.org/blog/700-million-and-premium#telegram-premium))

  - [x]  The maximum value of the field file_size in the classes [Animation](https://core.telegram.org/bots/api#animation), [Audio](https://core.telegram.org/bots/api#audio), [Document](https://core.telegram.org/bots/api#document), [Video](https://core.telegram.org/bots/api#video), [Voice](https://core.telegram.org/bots/api#voice), and [File](https://core.telegram.org/bots/api#file) can no longer be stored in a signed 32-bit integer type. This change is necessary to support 4GB files uploaded by [premium accounts](https://telegram.org/blog/700-million-and-premium#telegram-premium).
  - [x]  Added the field is_premium to the class [User](https://core.telegram.org/bots/api#user).
  - [x]  Added the field premium_animation to the class [Sticker](https://core.telegram.org/bots/api#sticker).

Attachment Menu Integration

  - [x]  Added the field added_to_attachment_menu to the class [User](https://core.telegram.org/bots/api#user).
  - [x]  Bots integrated in the attachment menu can now be used in groups, supergroups and channels.
  - [x]  Added support for t.me links that can be used to select the chat in which the attachment menu with the bot will be opened.

Other Changes

  - [x]  Added the parameter secret_token to the method [setWebhook](https://core.telegram.org/bots/api#setwebhook).
  - [x]  As previously announced, only HTTPS links are now allowed in login_url inline keyboard buttons.

+++ core/src/com/bot4s/telegram/models/Animation.scala
@@ -17,5 +17,5 @@ case class Animation(
   thumb: Option[PhotoSize] = None,
   fileName: Option[String] = None,
   mimeType: Option[String] = None,
-  fileSize: Option[Int] = None
+  fileSize: Option[Long] = None
 )
