+++ core/src/com/bot4s/telegram/models/Document.scala
@@ -16,5 +16,5 @@ case class Document(
   thumb: Option[PhotoSize] = None,
   fileName: Option[String] = None,
   mimeType: Option[String] = None,
-  fileSize: Option[Int] = None
+  fileSize: Option[Long] = None
 )
